+++ src/components/UpcomingConsultaionItem.tsx
@@ -1,0 +1,46 @@
+import { MdOutlineTimer } from 'react-icons/md';
+import Timer from './Timer';
+import { type TRequestedConsultationsProps } from '../lib/types';
+import { useNavigate } from 'react-router-dom';
+
+export const UpcomingConsultaionItem = ({
+  name,
+  title,
+  customer_id,
+  hopeDay,
+  hopeTime,
+  quick,
+}: TRequestedConsultationsProps) => {
+  const getBorderColorClass = (quick: boolean) => {
+    return quick ? 'quick-border' : 'border-gray-200';
+  };
+  const navigate = useNavigate();
+
+  const handleConsultationClick = (consultationId: number) => {
+    navigate(`/consulting/${consultationId}`);
+  };
+
+  return (
+    <div className={`bg-white rounded-lg p-4 mb-4 border ${getBorderColorClass(quick)} shadow-lg`}>
+      <div className='flex justify-between text-black text-[1rem] font-light'>
+        <span>{name} 손님</span>
+        <span>
+          {hopeDay} {hopeTime}
+        </span>
+      </div>
+      <div className='flex justify-between text-black text-[1rem] font-extrabold truncate mt-2'>
+        {title}
+        <span className='flex justify-center items-center gap-1'>
+          {quick && <MdOutlineTimer className='text-hanared' />}{' '}
+          {quick && <Timer hopeDay={hopeDay} hopeTime={hopeTime} />}{' '}
+        </span>
+        <button
+          className='border border-hanaindigo rounded-md px-1 text-[0.8rem] text-white bg-hanadeepgreen'
+          onClick={() => handleConsultationClick(customer_id)}
+        >
+          상담하기
+        </button>
+      </div>
+    </div>
+  );
+};
